# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: API do Observatório Turístico de Cumuruxatiba
  version: 0.7.0
  license:
    name: Licença a definir pelo Município de Prado
    identifier: LicenseRef-Proprietary
  description: >
    Contrato da fundação, estadias, questionários e métricas públicas. A
    extensão x-cumuru-feature informa a funcionalidade proprietária de cada
    operação — platform, core, questionnaire, analytics ou self-service — e o
    valor deferred marca a operação ainda não implementada. Tudo roda em modo
    PROTOTYPE_ONLY e somente com fixtures fictícias e generalizadas.
servers:
  - url: https://turismo.prado.ba.gov.br/api/v1
...
security:
  - oidc: []
paths:
...
  /accommodations/{accommodation_id}/calendar-feeds:
    get:
      summary: Listar feeds de calendário da acomodação
      description: >
        Devolve os feeds ativos e suspensos com o estado da última
        sincronização. A URL do feed nunca aparece na resposta: ela é segredo
        portador, e quem a tem lê o calendário do anúncio inteiro (ADR-044).
      tags: [calendar-feeds]
      operationId: listCalendarFeeds
      x-cumuru-feature: calendar-feed
      security:
        - oidc: ["stays:read:own"]
      parameters:
        - $ref: "#/components/parameters/AccommodationId"
      responses:
        "200":
          $ref: "#/components/responses/CalendarFeedList"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
    post:
      summary: Cadastrar feed de calendário
      description: >
        A hospedagem cola a URL .ics exportada do extranet da plataforma.
        Cadastrar um feed declara de onde vêm as datas daquela acomodação, que é
        o mesmo ato de quem registra estadia à mão, e por isso custa o mesmo
        escopo. Só https, sem credencial embutida no endereço e sem host
        interno; a URL é cifrada em repouso e não volta em nenhuma resposta.
      tags: [calendar-feeds]
      operationId: createCalendarFeed
      x-cumuru-feature: calendar-feed
      security:
        - oidc: ["stays:write"]
      parameters:
        - $ref: "#/components/parameters/AccommodationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarFeedRequest"
      responses:
        "201":
          $ref: "#/components/responses/CalendarFeedSucceeded"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/ProblemWithRetry"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
  /calendar-feeds/{feed_id}/remove:
    post:
      summary: Remover feed de calendário
      description: >
        Marca o feed como removido e para a sincronização. As estadias já
        confirmadas a partir dele continuam de pé: são fato da hospedagem, e
        apagar a origem apagaria a explicação de como entraram.
      tags: [calendar-feeds]
      operationId: removeCalendarFeed
      x-cumuru-feature: calendar-feed
      security:
        - oidc: ["stays:write"]
      parameters:
        - $ref: "#/components/parameters/CalendarFeedId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/EmptyObject"
      responses:
        "200":
          $ref: "#/components/responses/CalendarFeedSucceeded"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/ProblemWithRetry"
        "412":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
  /accommodations/{accommodation_id}/calendar-reservations:
    get:
      summary: Listar reservas importadas do calendário
      description: >
        Fila da acomodação, somando todos os feeds dela. Cada item é intenção de
        estadia observada na plataforma — datas e natureza — e nunca identidade
        ou número de hóspedes, porque o arquivo não traz nenhum dos dois.
      tags: [calendar-feeds]
      operationId: listCalendarReservations
      x-cumuru-feature: calendar-feed
      security:
        - oidc: ["stays:read:own"]
      parameters:
        - $ref: "#/components/parameters/AccommodationId"
        - $ref: "#/components/parameters/CalendarReservationStateFilter"
        - $ref: "#/components/parameters/CalendarPageLimit"
      responses:
        "200":
          $ref: "#/components/responses/CalendarReservationList"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
  /calendar-reservations/{reservation_id}/confirm:
    post:
      summary: Confirmar reserva importada e criar a estadia
      description: >
        Único caminho em que a observação vira presença. As datas vêm do
        calendário e o número de hóspedes vem de quem recebeu — o arquivo não o
        traz. A estadia é criada pelo mesmo caminho do cadastro à mão, na mesma
        transação, e por isso emite stay.created além da confirmação.
      tags: [calendar-feeds]
      operationId: confirmCalendarReservation
      x-cumuru-feature: calendar-feed
      security:
        - oidc: ["stays:write"]
      parameters:
        - $ref: "#/components/parameters/CalendarReservationId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmCalendarReservationRequest"
      responses:
        "200":
          $ref: "#/components/responses/CalendarReservationSucceeded"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/ProblemWithRetry"
        "412":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
  /calendar-reservations/{reservation_id}/dismiss:
    post:
      summary: Dispensar reserva importada
      description: >
        Diz que aquilo não era estadia: bloqueio de manutenção, uso do dono ou
        reserva que a hospedagem já registrou à mão. Não há motivo de lista
        fechada porque nada disso descreve pessoa alguma.
      tags: [calendar-feeds]
      operationId: dismissCalendarReservation
      x-cumuru-feature: calendar-feed
      security:
        - oidc: ["stays:write"]
      parameters:
        - $ref: "#/components/parameters/CalendarReservationId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/EmptyObject"
      responses:
        "200":
          $ref: "#/components/responses/CalendarReservationSucceeded"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/ProblemWithRetry"
        "412":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/InternalServerProblem"
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
    session:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Token opaco emitido por POST /auth/login. O servidor guarda somente o
        SHA-256 do token e o principal derivado tem os mesmos escopos de um
        principal OIDC.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
...
    AccommodationId:
      name: accommodation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CalendarFeedId:
      name: feed_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CalendarReservationId:
      name: reservation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CalendarReservationStateFilter:
      name: state
      in: query
      description: Recorte da fila; ausente devolve todos os estados.
      schema:
        $ref: "#/components/schemas/CalendarReservationState"
    CalendarPageLimit:
      name: limit
      in: query
      description: >
        A fila é limitada ao calendário de uma acomodação e não usa cursor: um
        token assinado numa lista que cabe na tela seria peso sem ganho.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  requestBodies:
...
    EmptyObject:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
  headers:
    RequestId:
      description: Identificador sanitizado de correlação da requisição.
      schema:
        type: string
    NoStore:
      description: Impede armazenamento de respostas operacionais.
      schema:
        type: string
        const: no-store
...
    EntityTag:
      description: ETag forte canônico correspondente ao version positivo.
      required: true
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
...
    IdempotencyReplayed:
      description: Informa se a resposta foi reconstruída de replay idempotente.
      required: true
      schema:
        type: string
        enum: ["true", "false"]
...
  responses:
...
    Problem:
      description: Erro no formato Problem Details.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalServerProblem:
      description: Erro interno sanitizado no formato Problem Details.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemWithRetry:
      description: Conflito; Retry-After aparece somente quando houver processamento concorrente.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          description: Presente somente para idempotency-in-progress.
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    CalendarFeedList:
      description: Feeds de calendário da acomodação, sem a URL de nenhum deles.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CalendarFeedList"
    CalendarFeedSucceeded:
      description: Comando aplicado ou reproduzido por replay idempotente.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        ETag:
          $ref: "#/components/headers/EntityTag"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CalendarFeed"
    CalendarReservationList:
      description: Fila de reservas importadas da acomodação.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CalendarReservationList"
    CalendarReservationSucceeded:
      description: Decisão aplicada ou reproduzida por replay idempotente.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        ETag:
          $ref: "#/components/headers/EntityTag"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CalendarReservation"
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        invalid_params:
          type: array
          items:
            $ref: "#/components/schemas/InvalidParam"
    InvalidParam:
      type: object
      required: [name, reason]
      additionalProperties: false
      properties:
        name:
          type: string
          description: JSON Pointer ou nome do campo inválido, sem o valor recebido.
        reason:
          type: string
          pattern: "^[a-z0-9-]+$"
...
    CalendarFeedProvider:
      type: string
      enum: [booking]
      description: >
        O provedor é campo, não abstração especulativa. Quando entrar o segundo,
        ele é uma linha nesta lista.
    CalendarFeedStatus:
      type: string
      enum: [active, suspended, removed]
      description: >
        Suspenso é o feed que falhou seguidamente e parou de ser buscado;
        removido continua na base porque as estadias confirmadas a partir dele
        continuam de pé.
    CalendarFeedSyncOutcome:
      type: string
      enum: [ok, unreachable, not_calendar, malformed]
      description: >
        not_calendar é o caso comum de URL expirada: o extranet redireciona para
        a tela de login em vez de falhar, e a hospedagem precisa ver isso
        distinto de um host fora do ar.
    CalendarFeed:
      type: object
      additionalProperties: false
      required:
        - id
        - accommodation_id
        - provider
        - label
        - status
        - last_synced_at
        - last_sync_outcome
        - consecutive_failures
        - version
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        accommodation_id:
          type: string
          format: uuid
        provider:
          $ref: "#/components/schemas/CalendarFeedProvider"
        label:
          type: string
          minLength: 1
          maxLength: 120
          description: Nome do anúncio dado pela hospedagem, nunca de pessoa.
        status:
          $ref: "#/components/schemas/CalendarFeedStatus"
        last_synced_at:
          type: [string, "null"]
          format: date-time
        last_sync_outcome:
          oneOf:
            - $ref: "#/components/schemas/CalendarFeedSyncOutcome"
            - type: "null"
        consecutive_failures:
          type: integer
          minimum: 0
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CalendarFeedList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CalendarFeed"
    CreateCalendarFeedRequest:
      type: object
      additionalProperties: false
      required: [provider, label, url]
      properties:
        provider:
          $ref: "#/components/schemas/CalendarFeedProvider"
        label:
          type: string
          minLength: 1
          maxLength: 120
        url:
          type: string
          format: uri
          maxLength: 2048
          pattern: "^https://"
          description: >
            Endereço .ics exportado do extranet. É segredo portador: entra por
            aqui, fica cifrado em repouso e não volta em nenhuma resposta.
    CalendarReservationKind:
      type: string
      enum: [reserved, blocked, unknown]
      description: >
        unknown é a resposta honesta para um resumo que ninguém escreveu para
        nós. O Booking.com não separa reserva de bloqueio de manutenção com
        confiabilidade, e adivinhar inflaria a presença publicada.
    CalendarReservationState:
      type: string
      enum: [pending, confirmed, dismissed, withdrawn]
      description: >
        withdrawn é a reserva que sumiu do feed. Sumiço não é cancelamento: a
        plataforma não diz que foi cancelada, apenas deixa de mostrá-la, e por
        isso a retirada nunca alcança o que já virou estadia.
    CalendarReservation:
      type: object
      additionalProperties: false
      required:
        - id
        - feed_id
        - arrival_on
        - departure_on
        - kind
        - state
        - stay_id
        - first_seen_at
        - last_seen_at
        - version
      properties:
        id:
          type: string
          format: uuid
        feed_id:
          type: string
          format: uuid
        arrival_on:
          type: string
          format: date
        departure_on:
          type: string
          format: date
        kind:
          $ref: "#/components/schemas/CalendarReservationKind"
        state:
          $ref: "#/components/schemas/CalendarReservationState"
        stay_id:
          type: [string, "null"]
          format: uuid
        first_seen_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 1
    CalendarReservationList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CalendarReservation"
    ConfirmCalendarReservationRequest:
      type: object
      additionalProperties: false
      required: [expected_guest_count, client_submission_id]
      properties:
        expected_guest_count:
          type: integer
          minimum: 1
          maximum: 100
          description: >
            Vem de quem recebeu os hóspedes. O calendário nunca traz este
            número, e é por isso que a confirmação é humana.
        client_submission_id:
          type: string
          format: uuid
          pattern: "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-7[0-9A-Fa-f]{3}-[89AaBb][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$"
...